--- templates/product-analytics/next/src/components/time-series-form.tsx
@@ -1,32 +1,0 @@
-import { eventTables } from "@/app/events";
-import MultiSelectForm from "./query-form/metric-select-form";
-import { ModelMeta } from "@/insights/model-meta";
-import { MetricForm } from "@/lib/form-types";
-import MetricSelectForm from "./metric-form";
-
-interface Props {
-  setForm: (form: MetricForm[]) => void;
-  setBreakdown: (form: string[]) => void;
-  breakdownOptions: ModelMeta[];
-}
-export default function TimeSeriesForm({
-  setForm,
-  breakdownOptions,
-  setBreakdown,
-}: Props) {
-  const list = eventTables.map((table) => ({
-    label: table.eventName,
-    val: table.tableName,
-  }));
-
-  const breakdownList = breakdownOptions.map((bd) => ({
-    label: bd.name,
-    val: bd.name,
-  }));
-
-  return (
-    <div>
-      <MetricSelectForm options={list} setForm={setForm} name="Metrics" />
-    </div>
-  );
-}
